introduce the concept of a model and collections of models of fields
Class tags for models are in the kanomi::model namespace.
Class tags for fields are in the kanomi::field namespace.

Introduce a Model<FIELD,STENCIL> template with a member typedef named
'list' that is a boost::fusion::list of types that implement the
field.

Rejigger the enabled and enabled_pred templates to work against model
collections.

Add a simple macro for enabling a particular implementation of a
model.

Figure out how to select the enabled implementation from the model
collection and append it to the evaluation sequence type.